refactor: :recycle: Added is_silent parameter to validation functions
Added is_silent parameter to `is_name_or_namespace_valid()`, `is_semver_valid()`, `validate_dependencies_and_incompatibilities()`, `is_mod_id_valid()`. Can be used to not log an error if the validation returns `false`. Usefull for example if the validaiton is used in a text input where the result will be most of the times `false`. Implemented with the ModTool in mind.