refactor: :recycle: Logging (#205)
* refactor: :recycle: created LogManager and LogEntry Class

Moved most of the logging logic from `ModLoaderUtils` into `LogManager`. New `LogEntry` Class to store log message info. Added log storing in `LogManager`.

works towards #140

* refactor: :recycle: Moved log functions to `ModLoaderLog`

Based on the discussion in #153 - Some compromises had to be done to prevent cyclic reference errors. No `ModLoaderUtils` functions could be used.

* refactor: :recycle: move `MOD_LOG_PATH` into *log.gd*

* refactor: :recycle: copied `ModLoaderUtils` into *mod_loader_setup.gd*

Had to copy the required utility functions into the setup script, because `ModLoaderUtils` depends now on other Classes that don't get loaded.

* refactor: :recycle: Moved `ModLoaderLogEntry` inside `ModLoaderLog`

With that the ModLoaderLog Class is fully independed and can be loaded in the setup script.

* fix: :pencil2: fixed search and replace errors

accidentally removed the `log_` from the deprecation message and `log_debug_json_print`.

* refactor: :recycle: created *setup_log.gd* and setup_utils.gd*

There tradeoffs for using the existing ModLoader Classes in the setup script are to big. So the required functions where moved into the *setup/* dir.

* refactor: :fire: clean up logged signal

* refactor: :recycle: Add `existing_entry` var

* feat: :sparkles: added time comparison

Added `time_stamp` to `ModLoaderLogEntry` and instead of saving the `time` string in `time_stamps` the same log entries are now stored in `stack`. Added `ModLoaderLogCompare` inner class, allowing to call the custom sort function.

* feat: :sparkles: Added `get_all_entries()` to `ModLoaderLogEntry`

* feat: :sparkles: added `get_all_as_resource_array()`

* feat: :sparkles: Added `get_all_as_string_array()`

* refactor: :fire: removed `time_stamps` in *setup_log.gd*

not needed in the setup logs

* fix: :bug: not using `.keys()`

* feat: :sparkles: Added `get_by_mod_as_resource_array()`

* feat: :sparkles: Added `get_by_mod_as_string_array()`

* feat: :sparkles: Added `get_by_type_as_resource_array()`

* feat: :sparkles: Added `get_by_type_as_string_array()`

* style: :pencil2: remove `_array` from func names

* refactor: :recycle: Added inner class `ModLoaderLogStoreUtils`

To prevent some of the duplications

* style: :art: direct return in `get_all_as_string()`

to match the other funcs

* refactor: :recycle: Use `ModLoaderLogStoreUtils` for all get funcs

* refactor: :recycle: Removed inner class `ModLoaderLogStoreUtils`

* fix: :pencil2: fixed accidentally remove func header

* refactor: :recycle: Updated *mod.gd* to use `ModLoaderLog`

* chore: :rewind: moving self setup changes to separate PR

* fix: :safety_vest: check if `.by_mod` has `mod_name`

* style: :pencil2: fix comments

Added description for `ignored_mod_names_in_log` in `ModLoaderStore` and removed the `;` in the `_rotate_log_file()` comment.

* style: :art: `get_entry()` using `str()` now