refactor: :recycle: simplified `get_mod_config()`
Only returning the full config data without the key parameter, as it only allowed retrieving top-level properties. | Removing the distinction between user config and default config. The default config is now generated based on the default key of the config_schema and saved to the mod config file. This ensures that there is always a user config. - To revert to the default config, simply delete the config JSON file.