refactor: :recycle: Renamed funcs | use a dict to store configs
Removed `mod` from functions in `config.gd` to maintain uniformity with other functions within this class. | Configurations are now stored inside a dictionary instead of an array in `ModData`, where the key represents the config name. This change makes it much easier to access and locate the appropriate config based on its name.