refactor: :recycle: only load configs if defaults are valid
`load_mod_config_defaults()` is now called in `load_configs()` in `mod_data.gd` to allow skipping the config loading if the default values in the schema are invalid. Additionally, the `default.json` file is now validated, so the default config is regenerated if the schema has changed and the defaults are no longer valid.