feat: ✨ Advanced Mod Configs (#237)
* feat: :sparkles: Added `JSONConfigFile` addon as submodule

* chore: :wrench: added `.gitmodules` to `.gitattributes`

* refactor: :recycle: remove `JSONConfigFile` add `JSON_Schema_Validator`

* chore: :arrow_up: update JSON Schema Validator

* feat: :sparkles: added `config_schema` to `ModManifest`

With it `_handle_mod_config()` | `_get_config_default_data()` generates the `config_default`s from the `config_schema` default keys | `get_config_default_data_as_string()` | `get_config_schema_as_string()` | `is_config_valid()` validates the config based on the schema using the `JSON_Schema_Validator` Addon

* feat: :sparkles: Save default conifg to File

If there is no json file for this mod config yet

* refactor: :recycle: load mod config after the manifest

With this change, the config data of the `config_defaults` defined in the `config_schema` is available in the first run.

* feat: :sparkles: added `get_dict_from_dict()`

used in `ModManifest` to get `config_schema`

* fix: :bug: only `_handle_mod_config()` if there is a schema

* refactor: :recycle: simplified `get_mod_config()`

Only returning the full config data without the key parameter, as it only allowed retrieving top-level properties. | Removing the distinction between user config and default config. The default config is now generated based on the default key of the config_schema and saved to the mod config file. This ensures that there is always a user config. - To revert to the default config, simply delete the config JSON file.

* refactor: :construction: Adding `ModConfig` Resource

* fix: :construction: mod config loading

remove `not` from if statement, and removed the call to the original `_load_mod_configs()` function

* chore: :bug:

* feat: :construction: wip: configs loading

* feat: :sparkles: added `get_dir_paths_in_dir()`

* feat: :sparkles: added `get_file_paths_in_dir()`

* style: :art: added missing empty line

* feat: :sparkles: imporved `ModConfig`

added `name` and `Is_valid` prop | Added `_init()` function to validate on init.

* refactor: :recycle: multiple entries per mod in `mod_list`

Added `current_config` to `mod_list` if the mod has a config file

* feat: :sparkles: added `get_mod_config_schema(mod_id: String)`

* feat: :sparkles: Implemented `current_config`

Reworked the config loading in *mod_manifest.gd* and *mod_data.gd* | The `current_config` for each mod is now stored in the user profiles.

* refactor: :recycle: Renamed funcs | use a dict to store configs

Removed `mod` from functions in `config.gd` to maintain uniformity with other functions within this class. | Configurations are now stored inside a dictionary instead of an array in `ModData`, where the key represents the config name. This change makes it much easier to access and locate the appropriate config based on its name.

* feat: :sparkles: added `save_to_disc()` in `ModConfig`

* feat: :sparkles: added `get_path_to_mod_configs()`

added `get_path_to_mod_configs()` in `_ModLoaderPath`

* feat: :sparkles: added `create_new_config()`

* refactor: :recycle: `set_current_config()` param changed to ModConfig

* feat: :sparkles: added `get_mods_with_config()`

* feat: :sparkles: added `current_config_changed(config)` signal

* feat: :sparkles: added `remove_file(file_path: String)`

and added missing types

* style: :art: renamed `create_new_config` -> `create_config`

also reworked error messages

* feat: :sparkles: added `delete_config()`

* feat: :sparkles: added `get_default_config()`

* feat: :sparkles: added `remove_from_disc()`

* refactor: :recycle: changed `_schema` param in `_init()`

default is an empty dictionary now, if empty the `config_schema` is retrieved from the `ModLoaderStore`. Otherwise, the provided schema in `_schema` is used.

* feat: :sparkles: added `get_schema_for_prop()`

* chore: :arrow_up: updated JSON_Schema_Validator

* refactor: :recycle: improved `schema_drill()` and renamed it

to `_traverse_schema()`

* feat: :sparkles: added `get_path_to_mod_config_file()`

* style: :pencil2: renamed `get_path_to_mod_configs()`

to `get_path_to_mod_configs_dir()`

* fix: :bug: update `mod_list` when loading profile

Fixed an error where the `current_config` specified in the `mod_list` of a user profile could point to a non-existent config if the config file was manually deleted.

* fix: :bug: Handle invalid `current_config` in a User Profile

If the `current_config` in a user profile is set to an invalid configuration, it will now fallback to the default configuration.

* refactor: :recycle: `create_config()` return `ModConfig`

Instead of returning a success boolean, `create_config()` now returns a `ModConfig` resource if successful.

* feat: :sparkles: added `update_config()`

* refactor: :recycle: update `is_valid` in `validate()`

* refactor: :recycle: updated `update_config(config: ModConfig)`

Using the `save_to_disc()` method of `ModConfig`

* refactor: :recycle: don't update config if `default`

* refactor: :recycle: added `DEFAULT_CONFIG_NAME`

added a const for the default config name

* refactor: :recycle: load **all** mod config files

Before, only valid configs were loaded into the `ModData`, which led to inconsistency between the stored config files and the configs in ModData. To prevent the need to check the file system all the time for existing configs, non-valid configs are now also added to the `ModData`.

* refactor: :recycle: use `get_default_config()`

* chore: :arrow_up: updated `JSON_Schema_Validator`

* style: :bulb: improved comments

* style: :bulb: added more comments

* style: :bulb: comment improvements

* style: :art: moved functions around

* style: :bulb: improved comments in *profile.gd*

* fix: :pencil2: fixed typo in comment

* fix: :bug: find and replace error

* style: :bulb: improved comments

* refactor: :recycle: inverted booleans in `is_valid()`

to make it more intuitive

* style: :art: renamed disk -> file

* style: :pencil2: updated comment to match function name

* refactor: :fire: removed unused `get_dir_paths_in_dir()`

* refactor: :recycle: only load configs if defaults are valid

`load_mod_config_defaults()` is now called in `load_configs()` in `mod_data.gd` to allow skipping the config loading if the default values in the schema are invalid. Additionally, the `default.json` file is now validated, so the default config is regenerated if the schema has changed and the defaults are no longer valid.

* feat: :sparkles: added `_ModLoaderCache`

Implemented `_ModLoaderCache` to store the MD5s of the config schema in `"user://ModLoaderCache.json"`. This enables checking if the config schema has changed since the last start, allowing the regeneration of the default config only if a change is detected.

* style: :bulb: added func comments for `_ModLoaderCache`

* refactor: :recycle: return cache data in `add_data()`

to align with the other methods

* refactor: :recycle: changed cache file name

* fix: :bug: handle non existing cache

Fixed an error that occurred when the `default.json` config file existed, but the cache had been deleted.