feat: :sparkles: option to allow autoload placement anywhere
To make things work when `ModLoaderStore` is not placed before `ModLoader`, I had to move the check to the beginning of the `_init()` function. To handle the problem in `check_autoload_order()`, where I couldn't use `ModLoaderLog.fatal()` due to its reliance on a function that references `ModLoaderStore`, I directly loaded the ModLoader Options resource in `_check_autoload_positions()`. Check out issue #262 for more info.

closes #257