refactor: ♻️ reworked mod uninstall detection (#298)
* feat: :sparkles: added ` get_dir_paths_in_dir()`

* feat: :sparkles: added `zip_name` and `zip_path` to `ModData`

* refactor: :recycle: added `zip_path` to user profiles

This allows to verify if the mod is still installed by confirming the existence of the zip file. However, this check is only performed when the mod is not loaded and a path to the zip file exists. This ensures that mods are not deleted from the profile when running in the editor. It's important to note that this new check may cause mods to appear in user profiles even if they are currently not loaded. To determine if a mod is actually loaded, you should also check `ModLoaderStore.mod_data` or use `ModLoaderMod.is_mod_loaded()`.

closes #288