refactor: :recycle: unique ids `optional_dependencies` - `dependencies`
Added validation to check for the same mod_ids in `optional_dependencies` and `dependencies`

Technically, it may break mods that have the same `mod_id` listed in both the `optional_dependencies` and `dependencies` sections. However, since the `optional_dependencies` feature was introduced in v6.0.0, it is highly unlikely that such conflicts exist.