refactor: :recycle: unique ids `load_before` - `incompatibilities`
Added validation to check for the same mod_ids in `load_before` and `incompatibilities`

Technically, it may break mods that have the same `mod_id` listed in both the `load_before` and `incompatibilities` sections. However, since the `load_before` feature was introduced in v6.0.0, it is highly unlikely that such conflicts exist.