fix: :bug: crash if no scene instance is returned (#392)
* fix: :bug: crash if no scene instance is returned

Fixes a crash where the `scene_instance` is not returned by an `edit_callable`. Added a check for the existence of `edited_scene` before invoking `_save_scene()`. If the `scene_instance` hasn't been returned by the callable, an appropriate error message is logged. Additionally, a `SceneEditCallable` subclass is added to `_ModLoaderSceneExtension` to store the `mod_id` responsible for modifying the scene, enhancing the specificity of the error message.

* docs: :bulb: added missing `mod_id` param comment

* refactor: :fire: removed `mod_id` param from `extend_scene`

also removed the `SceneEditCallable` subclass from `_ModLoaderSceneExtension`.

* refactor: :fire: Removed no longer accurate comment